+++ Sources/Device.swift
@@ -19,7 +19,7 @@
 
 public extension Device {
     /// The version of the Device Library since cannot get directly from Package.
-    static let version: Version = "2.10.19"
+    static let version: Version = "2.11.0"
 }
 import Compatibility
 
@@ -423,8 +423,8 @@ public struct Device: IdiomType, Hashable, CustomStringConvertible, Identifiable
                 return [.headphoneJack, .screen(.tv)]
             case .watch:
                 return [.battery, .wirelessCharging, .nfc, .applePay, .roundedCorners]
-            case .vision:
-                return [.battery, .biometrics(.opticID), .lidar, .cameras([.stereoscopic, .persona]), .screen(.p720), .appleIntelligence]
+            case .vision: // All visions are pro for now.  When this is no longer the case, move this to each device.
+                return [.pro, .battery, .biometrics(.opticID), .lidar, .cameras([.stereoscopic, .persona]), .screen(.p720), .appleIntelligence]
             case .homePod:
                 return [.screen(.w38)]
             case .unspecified, .mac, .carPlay: